+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [337] - 2026-03-04
+
+- [TASK-483] Render 7x24 day-of-week hour activity heatmap in Cost tab
+
 ## [336] - 2026-03-04
 
 - [TASK-482] Render hour-of-day cost bar chart in Dashboard Cost tab